app/source/postgresdb/requirements.txt
Faker==22.5.0
psycopg2-binary==2.9.3
